Move build constants to `buildSrc`
