+++ buildSrc/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins { 
+    `kotlin-dsl`
+}
+
+repositories {
+    jcenter()
+}
