+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,0 +1,20 @@
+@Suppress("ObjectPropertyName")
+object Versions {
+    const val kotlin = "1.3.30"
+    const val scalaBinary = "2.12"
+    const val lagom = "1.5.1" // "1.4.13"
+    const val ktlint = "0.32.0"
+    const val `ktlint-plugin` = "8.0.0"
+    const val `kotlin-logging` = "1.6.10"
+    const val config4k = "0.4.1"
+    const val `lagom-extensions` = "0.1.0"
+    const val swagger = "2.0.7"
+    const val jacoco = "0.8.2"
+    const val junit5 = "5.3.2"
+    const val assertj = "3.11.1"
+    const val `json-unit` = "2.6.1"
+    const val `nexus-staging` = "0.20.0"
+    const val `nexus-publish` = "0.2.0"
+    const val dokka = "0.9.18"
+    const val lombok = "3.5.1"
+}
